Fixed bug in open_file function
There was a small issue where we were not stripping the path correctly.
Once this is resolved, the files are opended correctly.  We still retain
the functionality that we respond with the path. That should be changed
to a 404 though instead, since it makes more sense for my static file
server.